Renamed SClass to more accurately reflect use
